LayoutTests/storage/indexeddb/modern/idbcursor-continue-primary-key-1-expected.txt

This test verifies the basic functionality of IDBCursor.continuePrimaryKey().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !!cursor is true
PASS "transaction-state check should precede deletion check" threw as expected
PASS !!cursor is true
PASS "deletion check should precede index source check" threw as expected
PASS !!cursor is true
PASS "deletion check should precede cursor direction check" threw as expected
PASS "direction check should precede got_value_flag check" threw as expected
PASS !!cursor is true
PASS "direction check should precede iteration ongoing check" threw as expected
PASS !!cursor is true
PASS "index source check should precede iteration ongoing check" threw as expected
PASS "index source check should precede got_value_flag check" threw as expected
PASS !!cursor is true
PASS "iteration ongoing check should precede unset key check" threw as expected
PASS "got_value_flag check should precede unset key check" threw as expected
PASS !!cursor is true
PASS "DataError is expected if key is unset." threw as expected
PASS !!cursor is true
PASS "DataError is expected if primary key is unset." threw as expected
PASS !!cursor is true
PASS "DataError is expected if key is lower then current one." threw as expected
PASS "DataError is expected if primary key is equal to current one." threw as expected
PASS "DataError is expected if primary key is lower than current one." threw as expected
PASS !!cursor is true
PASS "DataError is expected if key is larger then current one." threw as expected
PASS "DataError is expected if primary key is equal to current one." threw as expected
PASS "DataError is expected if primary key is larger than current one." threw as expected
PASS successfullyParsed is true

TEST COMPLETE
